--- app/src/main/java/com/example/wellcome/BookmarkFragment.kt
@@ -1,4 +1,0 @@
-package com.example.wellcome
-
-class BookmarkFragment {
-}
