+++ app/src/main/java/com/example/wellcome/community/CommunityFragment.kt
@@ -1,21 +1,21 @@
-package com.example.wellcome
+package com.example.wellcome.community
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.wellcome.databinding.FragmentTravelBinding
+import com.example.wellcome.databinding.FragmentCommunityBinding
 
-class TravelFragment : Fragment() {
-
-    lateinit var binding: FragmentTravelBinding
+class CommunityFragment : Fragment() {
+    lateinit var binding: FragmentCommunityBinding
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentTravelBinding.inflate(inflater, container, false)
+        binding = FragmentCommunityBinding.inflate(inflater, container, false)
         return binding.root
     }
+
 }
