+++ app/src/main/java/com/example/wellcome/main/ArticleAdapter.kt
@@ -1,13 +1,7 @@
-package com.example.wellcome
+package com.example.wellcome.main
 
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.TextView
-import androidx.cardview.widget.CardView
-import androidx.recyclerview.widget.DiffUtil
-import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.example.wellcome.databinding.ItemArticleBinding
