+++ app/src/main/res/drawable/img_gear.xml
