--- app/src/main/res/drawable/setting.xml
@@ -1,11 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-  <path
-      android:pathData="M16,12C16,14.209 14.21,16 12,16C9.791,16 8,14.209 8,12C8,9.791 9.791,8 12,8C14.21,8 16,9.791 16,12Z"
-      android:strokeWidth="2"
-      android:fillColor="#00000000"
-      android:strokeColor="#000000"/>
-</vector>
