+++ app/src/main/res/drawable/style_article.xml
