+++ app/src/main/res/layout/fragment_main.xml
@@ -20,7 +20,7 @@
             android:layout_height="40dp"
             android:layout_gravity="center_vertical"
             android:layout_marginRight="8dp"
-            android:src="@drawable/logo" />
+            android:src="@drawable/img_logo" />
 
         <TextView
             android:id="@+id/profileText"
@@ -36,7 +36,7 @@
             android:layout_width="24dp"
             android:layout_height="24dp"
             android:layout_gravity="center"
-            android:src="@drawable/bell"
+            android:src="@drawable/img_bell"
             android:background="@color/white"/>
 
         <ImageButton
@@ -44,7 +44,7 @@
             android:layout_width="24dp"
             android:layout_height="24dp"
             android:layout_gravity="center"
-            android:src="@drawable/gear"
+            android:src="@drawable/img_gear"
             android:layout_marginStart="8dp"
             android:background="@color/white"/>
 
@@ -74,9 +74,9 @@
         android:paddingHorizontal="24dp"
         android:paddingVertical="16dp"
         android:layout_marginVertical="32dp"
-        android:background="@drawable/button_style"
+        android:background="@drawable/style_button"
         android:backgroundTint="@color/main_color"
-        android:drawableLeft="@drawable/leaf"
+        android:drawableLeft="@drawable/img_leaf"
         android:text="여행지 추천받기"
         android:textSize="18sp"
         android:fontFamily="@font/pretendard_bold"/>
