+++ app/src/main/res/layout/item_article.xml
@@ -13,14 +13,14 @@
             android:id="@+id/article_img"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:src="@drawable/temple"
+            android:src="@drawable/img_temple"
             android:scaleType="centerCrop" />
         <LinearLayout
             android:id="@+id/article_text_layout"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="vertical"
-            android:background="@drawable/article_style"
+            android:background="@drawable/style_article"
             android:padding="20dp"
             android:layout_gravity="bottom">
             <TextView
