+++ app/src/main/res/menu/menu_bottom.xml
@@ -6,19 +6,19 @@
         android:id="@+id/tab1"
         app:showAsAction="always"
         android:enabled="true"
-        android:icon="@drawable/home"
+        android:icon="@drawable/img_home"
         android:title="메인" />
     <item
         android:id="@+id/tab2"
         app:showAsAction="always"
         android:enabled="true"
-        android:icon="@drawable/alt"
+        android:icon="@drawable/img_bookmark"
         android:title="여행일정" />
     <item
         android:id="@+id/tab3"
         app:showAsAction="always"
         android:enabled="true"
-        android:icon="@drawable/user"
+        android:icon="@drawable/img_user"
         android:title="커뮤니티" />
 
 </menu>
